requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
bcrypt==4.0.1
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
dnspython==2.6.1
email-validator==1.3.1
exceptiongroup==1.2.2
fastapi==0.89.1
fastapi-users==10.3.0
fastapi-users-db-sqlalchemy==4.0.4
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.8
makefun==1.15.4
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
passlib==1.7.4
pycparser==2.22
pydantic==1.10.4
pydantic_core==2.20.1
PyJWT==2.6.0
python-dotenv==1.0.1
python-multipart==0.0.5
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.22.0
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
yarl==1.9.4
